django_service/requirements.txt
asgiref==3.8.1
Django==3.2.16
django-cors-headers==3.13.0
django-redis==5.4.0
djangorestframework==3.12.4
drf-yasg==1.21.10
inflection==0.5.1
kafka-python==2.2.11
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
uritemplate==4.2.0
